add two new files, split CharacterList and Character into separate class component file
