+++ starwars/src/App.css
@@ -30,8 +30,8 @@ ul {
   font-size: 32px;
 } */
 
-.grow {transition: all .4s ease-in-out;}
-.grow:hover {transform: scale(1.5);}
+.grow {transition: all .3s ease-in-out;}
+.grow:hover {transform: scale(1.3);}
 
 .main-content {
     -webkit-perspective: 300;
