+++ starwars/src/App.js
@@ -1,5 +1,7 @@
 import React, { Component } from 'react';
 import './App.css';
+import CharacterList from './components/CharacterList'
+import Character from './components/Character'
 
 class App extends Component {
   constructor() {
@@ -51,53 +53,53 @@ class App extends Component {
 }
 
 
-class CharacterList extends React.Component {
-  render() {
-    return this.props.grabResultsArray.map((e) => {
-      return (
-        <div>
-            <ul>
-                <li>
-                    <a> 
-                    <Character key={e.created} 
-                      characterName={e.name} 
-                      birthdate={e.birth_year} 
-                      gender={e.gender}
-                      height={e.height}
-                      mass={e.mass}
-                      eye={e.eye_color}
-                      hair={e.hair_color}
-                      skin={e.skin_color}
+// class CharacterList extends React.Component {
+//   render() {
+//     return this.props.grabResultsArray.map((e) => {
+//       return (
+//         <div>
+//             <ul>
+//                 <li>
+//                     <a> 
+//                     <Character key={e.created} 
+//                       characterName={e.name} 
+//                       birthdate={e.birth_year} 
+//                       gender={e.gender}
+//                       height={e.height}
+//                       mass={e.mass}
+//                       eye={e.eye_color}
+//                       hair={e.hair_color}
+//                       skin={e.skin_color}
                       
-                    />
-                    </a>
-                </li>
-            </ul>
-        </div>
-      )
-    });
-  }
-}
+//                     />
+//                     </a>
+//                 </li>
+//             </ul>
+//         </div>
+//       )
+//     });
+//   }
+// }
 
-class Character extends React.Component {
-  render() {
-    return (
-      <div className="grow">
-          <article>
-            <p> Name: {this.props.characterName},  
-                Born: {this.props.birthdate},
-                Gender: {this.props.gender},
-                Height: {this.props.height},
-                Mass: {this.props.mass},
-                Eye: {this.props.eye},
-                Hair: {this.props.hair},
-                Skin: {this.props.skin},
-            </p>
+// class Character extends React.Component {
+//   render() {
+//     return (
+//       <div className="grow">
+//           <article>
+//             <p> Name: {this.props.characterName},  
+//                 Born: {this.props.birthdate},
+//                 Gender: {this.props.gender},
+//                 Height: {this.props.height},
+//                 Mass: {this.props.mass},
+//                 Eye: {this.props.eye},
+//                 Hair: {this.props.hair},
+//                 Skin: {this.props.skin},
+//             </p>
             
-          </article>
-      </div>
-    )
-  }
-}
+//           </article>
+//       </div>
+//     )
+//   }
+// }
 
 export default App;
